added stride computation for all layout
need to add unit test
